Outlets fix

+++ app/routes/outlets/outletsCRUD.js
@@ -1,6 +1,8 @@
 const createModelMiddleware = require('../middleware/createModelMiddleware');
 const findModelByMiddleware = require('../middleware/findModelByMiddleware');
+const checkEndpointAccessMiddleware = require('../../auth/middleware/checkEndpointAccessMiddleware');
 const paginateMiddleware = require('../middleware/paginateMiddleware');
+const authMiddleware = require('../../auth/middleware/authMiddleware');
 const {OutletModel} = require('../../models/appModels');
 
 module.exports = (router) => {
@@ -12,18 +14,24 @@ module.exports = (router) => {
         });
 
     router.get('/outlets/:id',
+        authMiddleware(),
+        checkEndpointAccessMiddleware(4),
         findModelByMiddleware(OutletModel, {_id: 'id'}, 'params', 'outlet'),
         (req, res) => {
             res.jsonSuccess(req.outlet);
         });
 
     router.post('/outlets',
+        authMiddleware(),
+        checkEndpointAccessMiddleware(4),
         createModelMiddleware(OutletModel, 'outlet'),
         (req, res) => {
             res.jsonSuccess(req.outlet);
         });
 
     router.patch('/outlets/:id',
+        authMiddleware(),
+        checkEndpointAccessMiddleware(4),
         findModelByMiddleware(OutletModel, {_id: 'id'}, 'params', 'outlet'),
         async (req, res) => {
             try {
@@ -38,6 +46,8 @@ module.exports = (router) => {
         });
 
     router.delete('/outlets/:id',
+        authMiddleware(),
+        checkEndpointAccessMiddleware(4),
         findModelByMiddleware(OutletModel, {_id: 'id'}, 'req.params', 'outlet'),
         async (req, res) => {
             await req.outlet.delete();
